requirements.txt
torchsummary~=1.5
numpy~=1.24.3
pillow~=10.4.0
torchvision~=0.11.2
pandas~=1.3.4
matplotlib~=3.7.5